fix(dashboard,api-service): remember agent provider during onboarding fixes NV-7416
